+++ src/components/Instructor.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import styles from './Instructor.module.css';
+import styles from './styles/Instructor.module.css';
 
 const Instructor = props => {
   const { instructor } = props;
