+++ src/components/UserSchedule.js
@@ -2,9 +2,9 @@ import React from 'react';
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import { Link, Redirect } from 'react-router-dom';
-import { CreateUserRows, getDay } from '../../helpers/CalendarHelper';
-import { deleteClass } from '../../services/classes';
-import '../InstructorSchedule/InstructorSchedule.module.css';
+import { CreateUserRows, getDay } from '../helpers/CalendarHelper';
+import { deleteClass } from '../services/classes';
+import './styles/InstructorSchedule.module.css';
 
 const UserSchedule = props => {
   const {
