+++ src/components/styles/Home.module.css
@@ -5,6 +5,7 @@
 
 .gym_image {
   width: 80%;
+  height: 30%;
 }
 
 .inst_button_container {
