+++ src/components/styles/InstructorSchedule.module.css
