+++ src/containers/Header.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Header from '../components/Header/Header';
+import Header from '../components/Header';
 import { logout } from '../actions';
 
 const mapStateToProps = state => ({ user: state.user });
