+++ src/containers/InstructorFile.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { changeWeek } from '../actions';
-import InstructorFile from '../components/InstructorFile/InstructorFile';
+import InstructorFile from '../components/InstructorFile';
 
 const mapStateToProps = state => ({
   instructors: state.instructor,
