+++ src/containers/Login.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Login from '../components/Login/Login';
+import Login from '../components/Login';
 import { login, addUsers } from '../actions';
 
 const mapStateToProps = state => ({ user: state.user, registeredUsers: state.registeredUsers });
